fix time to player lose and fix money bet, time to show big win

+++ core/src/com/ss/repository/Leaderboard.java
@@ -64,7 +64,7 @@ public class Leaderboard extends GScreen {
 
   public interface RankingService {
     int getNRank();
-    int getCurrentScore(int rank);
+    long getCurrentScore(int rank);
   }
 
   private                   TextureAtlas                      lbAtlas;
